Test | CAKK-36 | 로직 리팩토링으로 인한 테스트 리팩토링
