+++ cakk-domain/mysql/src/test/java/com/cakk/domain/base/DomainTest.java
@@ -11,11 +11,16 @@ import com.navercorp.fixturemonkey.FixtureMonkey;
 import com.navercorp.fixturemonkey.api.introspector.BuilderArbitraryIntrospector;
 import com.navercorp.fixturemonkey.api.introspector.ConstructorPropertiesArbitraryIntrospector;
 import com.navercorp.fixturemonkey.api.introspector.FieldReflectionArbitraryIntrospector;
+import com.navercorp.fixturemonkey.customizer.Values;
 import com.navercorp.fixturemonkey.jakarta.validation.plugin.JakartaValidationPlugin;
 
 import com.cakk.common.enums.Role;
+import com.cakk.common.enums.VerificationStatus;
 import com.cakk.domain.mysql.bo.user.DefaultVerificationPolicy;
 import com.cakk.domain.mysql.bo.user.VerificationPolicy;
+import com.cakk.domain.mysql.dto.param.user.CertificationParam;
+import com.cakk.domain.mysql.entity.shop.CakeShop;
+import com.cakk.domain.mysql.entity.user.BusinessInformation;
 import com.cakk.domain.mysql.entity.user.User;
 
 public abstract class DomainTest {
@@ -62,5 +67,37 @@ public abstract class DomainTest {
 
 	protected VerificationPolicy getVerificationPolicy() {
 		return new DefaultVerificationPolicy();
+	}
+
+	protected CakeShop getCakeShopFixture() {
+		return getConstructorMonkey().giveMeBuilder(CakeShop.class)
+			.set("shopName", Arbitraries.strings().withCharRange('a', 'z').ofMaxLength(30))
+			.set("shopBio", Arbitraries.strings().withCharRange('a', 'z').ofMaxLength(40))
+			.set("shopDescription", Arbitraries.strings().withCharRange('a', 'z').ofMaxLength(500))
+			.set("location", supplyPointBy(
+				Arbitraries.doubles().between(-90, 90).sample(),
+				Arbitraries.doubles().between(-180, 180).sample())
+			)
+			.sample();
+	}
+
+	protected BusinessInformation getBusinessInformationFixtureWithCakeShop(VerificationStatus verificationStatus) {
+		return getConstructorMonkey().giveMeBuilder(BusinessInformation.class)
+			.set("businessNumber", Arbitraries.strings().withCharRange('a', 'z').ofMaxLength(20))
+			.set("cakeShop", Values.just(getCakeShopFixture()))
+			.set("verificationStatus", verificationStatus)
+			.setNull("user")
+			.sample();
+	}
+
+	protected CertificationParam getCertificationParamFixtureWithUser(User user) {
+		return getBuilderMonkey().giveMeBuilder(CertificationParam.class)
+			.set("businessRegistrationImageUrl", Arbitraries.strings().withCharRange('a', 'z').ofMinLength(1).ofMaxLength(20))
+			.set("idCardImageUrl", Arbitraries.strings().withCharRange('a', 'z').ofMinLength(1).ofMaxLength(20))
+			.set("cakeShopId", Arbitraries.longs().greaterOrEqual(0))
+			.set("emergencyContact", Arbitraries.strings().withCharRange('a', 'z').ofMinLength(1).ofMaxLength(20))
+			.set("message", Arbitraries.strings().withCharRange('a', 'z').ofMaxLength(20))
+			.set("user", user)
+			.sample();
 	}
 }
