+++ cakk-domain/mysql/src/test/java/com/cakk/domain/entity/user/VerificationPolicyTest.java
@@ -2,12 +2,18 @@ package com.cakk.domain.entity.user;
 
 import static org.assertj.core.api.Assertions.*;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import com.cakk.common.enums.Role;
 import com.cakk.common.enums.VerificationStatus;
+import com.cakk.common.exception.CakkException;
 import com.cakk.domain.base.DomainTest;
 import com.cakk.domain.mysql.bo.user.VerificationPolicy;
+import com.cakk.domain.mysql.dto.param.user.CertificationParam;
+import com.cakk.domain.mysql.entity.user.BusinessInformation;
+import com.cakk.domain.mysql.entity.user.User;
 
 public class VerificationPolicyTest extends DomainTest {
 
@@ -27,6 +33,48 @@ public class VerificationPolicyTest extends DomainTest {
 		VerificationPolicy verificationPolicy = getVerificationPolicy();
 
 		assertThat(verificationPolicy.isCandidate(VerificationStatus.PENDING)).isTrue();
+	}
+
+	@Test
+	@DisplayName("사장님 인증된 케이크 샵이 존재할 때, 인증 요청에 실패한다")
+	void registerCertificationInformation2() {
+		//given
+		VerificationPolicy verificationPolicy = getVerificationPolicy();
+		BusinessInformation businessInformation = getBusinessInformationFixtureWithCakeShop(VerificationStatus.APPROVED);
+		User user = getUserFixture(Role.USER);
+		CertificationParam param = getCertificationParamFixtureWithUser(user);
+
+		//when, then
+		Assertions.assertThrowsExactly(CakkException.class,
+			() -> verificationPolicy.requestCertificationBusinessOwner(businessInformation, param));
+	}
+
+	@Test
+	@DisplayName("사장님 인증 요청이 진행중일 때, 인증 요청에 실패한다")
+	void registerCertificationInformation3() {
+		//given
+		VerificationPolicy verificationPolicy = getVerificationPolicy();
+		BusinessInformation businessInformation = getBusinessInformationFixtureWithCakeShop(VerificationStatus.PENDING);
+		User user = getUserFixture(Role.USER);
+		CertificationParam param = getCertificationParamFixtureWithUser(user);
+
+		//when, then
+		Assertions.assertThrowsExactly(CakkException.class,
+			() -> verificationPolicy.requestCertificationBusinessOwner(businessInformation, param));
+	}
+
+	@Test
+	@DisplayName("사장님 인증 요청이 거절 됐을 때, 인증 요청에 실패한다")
+	void registerCertificationInformation4() {
+		//given
+		VerificationPolicy verificationPolicy = getVerificationPolicy();
+		BusinessInformation businessInformation = getBusinessInformationFixtureWithCakeShop(VerificationStatus.REJECTED);
+		User user = getUserFixture(Role.USER);
+		CertificationParam param = getCertificationParamFixtureWithUser(user);
+
+		//when, then
+		Assertions.assertThrowsExactly(CakkException.class,
+			() -> verificationPolicy.requestCertificationBusinessOwner(businessInformation, param));
 	}
 }
